chore: cleaning file structure for action card
