--- packages/react/src/components/ActionCard/ActionCard.styles.ts
@@ -1,7 +1,0 @@
-import { card, styled } from '@project44-manifest/react-styles';
-
-export const StyledActionCard = styled('a', card, {
-  display: 'inline-block',
-  color: '$text-primary',
-  textDecoration: 'none',
-});
