--- packages/react/src/components/ActionCard/index.ts
@@ -1,2 +1,0 @@
-export { ActionCard } from './ActionCard';
-export type { ActionCardElement, ActionCardProps } from './ActionCard.types';
