--- packages/react/src/components/ActionCardBody/ActionCardBody.styles.ts
@@ -1,8 +1,0 @@
-import { styled } from '@project44-manifest/react-styles';
-
-export const StyledActionCardBody = styled('div', {
-  px: '$large',
-  paddingBottom: '$large',
-  typography: '$body',
-  width: '100%',
-});
