--- packages/react/src/components/ActionCardBody/ActionCardBody.types.ts
@@ -1,10 +1,0 @@
-import type { CSS } from '@project44-manifest/react-styles';
-
-export type ActionCardBodyElement = 'div';
-
-export interface ActionCardBodyProps {
-  /**
-   * Theme aware style object
-   */
-  css?: CSS;
-}
