--- packages/react/src/components/ActionCardBody/index.ts
@@ -1,2 +1,0 @@
-export { ActionCardBody } from './ActionCardBody';
-export type { ActionCardBodyElement, ActionCardBodyProps } from './ActionCardBody.types';
