--- packages/react/src/components/ActionCardHeader/ActionCardHeader.styles.ts
@@ -1,10 +1,0 @@
-import { styled } from '@project44-manifest/react-styles';
-
-export const StyledActionCardHeader = styled('div', {
-  alignItems: 'center',
-  display: 'flex',
-  justifyContent: 'space-between',
-  width: '100%',
-  p: '$large',
-  typography: '$title',
-});
