--- packages/react/src/components/ActionCardHeader/ActionCardHeader.types.ts
@@ -1,10 +1,0 @@
-import type { CSS } from '@project44-manifest/react-styles';
-
-export type ActionCardHeaderElement = 'div';
-
-export interface ActionCardHeaderProps {
-  /**
-   * Theme aware style object
-   */
-  css?: CSS;
-}
