--- packages/react/src/components/ActionCardHeader/index.ts
@@ -1,2 +1,0 @@
-export { ActionCardHeader } from './ActionCardHeader';
-export type { ActionCardHeaderElement, ActionCardHeaderProps } from './ActionCardHeader.types';
