--- packages/react/src/components/ActionCardImage/index.ts
@@ -1,2 +1,0 @@
-export { ActionCardImage } from './ActionCardImage';
-export type { ActionCardImageElement, ActionCardImageProps } from './ActionCardImage.types';
