+++ packages/react/src/components/action-card/index.ts
@@ -1,0 +1,1 @@
+export * from './action-card';
